Update existing github comment

+++ src/yaml_lint.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-function yaml_lint {
+yaml_lint() {
 
     # gather output
     echo "lint: info: yamllint on ${yamllint_file_or_dir}."
@@ -25,7 +25,11 @@ function yaml_lint {
 
     # comment if lint failed
     if [ "${GITHUB_EVENT_NAME}" == "pull_request" ] && [ "${yamllint_comment}" == "1" ] && [ ${lint_exit_code} -ne 0 ]; then
-        lint_comment_wrapper="#### \`yamllint\` ${lint_comment_status}
+        header="<!-- yamllint -->"
+
+        lint_comment_wrapper="
+${header}
+#### \`yamllint\` ${lint_comment_status}
 <details><summary>Show Output</summary>
 
 \`\`\`
@@ -34,15 +38,21 @@ ${lint_output}
 </details>
 
 *Workflow: \`${GITHUB_WORKFLOW}\`, Action: \`${GITHUB_ACTION}\`, Lint: \`${yamllint_file_or_dir}\`*"
-    
+
         echo "lint: info: creating json"
         lint_payload=$(echo "${lint_comment_wrapper}" | jq -R --slurp '{body: .}')
         lint_comment_url=$(cat ${GITHUB_EVENT_PATH} | jq -r .pull_request.comments_url)
         echo "lint: info: commenting on the pull request"
-        echo "${lint_payload}" | curl -s -S -H "Authorization: token ${GITHUB_ACCESS_TOKEN}" --header "Content-Type: application/json" --data @- "${lint_comment_url}" > /dev/null
+        existing_comment=$(get_comments | jq '[.[] | select(.body | startswith("'"${header}"'"))] | first | .id')
+        if [ "${existing_comment}" = "null" ]; then
+            # new comment
+            echo "${lint_payload}" | curl -s -S -H "Authorization: token ${GITHUB_ACCESS_TOKEN}" --header "Content-Type: application/json" --data @- "${lint_comment_url}" > /dev/null
+        else
+            # update existing comment
+            echo "${lint_payload}" | curl -s -S -H "Authorization: token ${GITHUB_ACCESS_TOKEN}" -X PATCH --header "Content-Type: application/json" --data @- "${lint_comment_url}/${existing_comment}" > /dev/null
+        fi
     fi
 
     echo ::set-output name=yamllint_output::${lint_output}
     exit ${lint_exit_code}
 }
-
